+++ src/main/java/Duke.java
@@ -1,3 +1,8 @@
+import exceptions.DukeException;
+import handlers.DukeCommand;
+import handlers.DukeCommandMap;
+import models.Task;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
