+++ src/main/java/exceptions/DukeException.java
@@ -1,3 +1,5 @@
+package exceptions;
+
 public class DukeException extends Exception{
     public String errorMessage;
