+++ src/main/java/handlers/AddTodoCommand.java
@@ -1,8 +1,15 @@
-import java.util.List;
+package handlers;
 
-public class AddTodoCommand implements DukeCommand{
+import exceptions.DukeException;
+import models.Task;
+import models.Todo;
+
+import java.util.List;
+import java.util.Objects;
+
+public class AddTodoCommand implements DukeCommand {
     public String run (List<Task> taskList, String s) throws DukeException {
-        if (s == "") {
+        if (Objects.equals(s, "")) {
             throw new DukeException("Todo must not be empty!\n");
         }
         Todo todo = new Todo(s);
