+++ src/main/java/handlers/DeleteTaskCommand.java
@@ -1,11 +1,14 @@
+package handlers;
+
+import models.Task;
+
 import java.util.List;
 
-public class MarkAsDoneCommand implements DukeCommand {
+public class DeleteTaskCommand implements DukeCommand {
     public String run (List<Task> taskList, String content) {
         try {
-            Integer taskNum = Integer.valueOf(content);
-            taskList.get(taskNum - 1).markAsDone(); // since taskList is 0-indexed
-            return "Good job on completing the task:\n" + taskList.get(taskNum - 1);
+            int taskNum = Integer.parseInt(content);
+            return "Task deleted:\n" + taskList.remove(taskNum - 1);
         } catch (NumberFormatException e) {
             return "Task index is not a number!\n";
         } catch (IndexOutOfBoundsException e) {
