+++ src/main/java/handlers/ListTasksCommand.java
@@ -1,10 +1,15 @@
+package handlers;
+
+import exceptions.DukeException;
+import models.Task;
+
 import java.util.List;
 
 public class ListTasksCommand implements DukeCommand {
 
-    public static String listTask(List<Task> taskList) {
+    public String run (List<Task> taskList, String content) {
         if (taskList.size() == 0) {
-            return "No tasks added yet!\n";
+             return "No tasks added yet!\n";
         }
 
         StringBuilder fullList = new StringBuilder();
@@ -12,9 +17,5 @@ public class ListTasksCommand implements DukeCommand {
             fullList.append(i + 1).append(". ").append(taskList.get(i));
         }
         return fullList.toString();
-    }
-
-    public String run (List<Task> taskList, String content) {
-        return listTask(taskList);
     }
 }
