+++ src/main/java/handlers/MarkAsUndoneCommand.java
@@ -1,9 +1,13 @@
+package handlers;
+
+import models.Task;
+
 import java.util.List;
 
 public class MarkAsUndoneCommand implements DukeCommand {
     public String run (List<Task> taskList, String content) {
         try {
-            Integer taskNum = Integer.valueOf(content);
+            int taskNum = Integer.parseInt(content);
             taskList.get(taskNum - 1).markAsUndone();
             return "This task is not done yet:\n" + taskList.get(taskNum - 1);
         } catch (NumberFormatException e) {
