+++ src/main/java/models/Deadline.java
@@ -1,4 +1,8 @@
-public class Deadline extends Task{
+package models;
+
+import models.Task;
+
+public class Deadline extends Task {
     protected String by;
 
     public Deadline(String description, String by) {
