+++ src/main/java/models/Todo.java
@@ -1,3 +1,5 @@
+package models;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description);
